@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.4.2</version>
+		<version>3.1.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.springboot.blog</groupId>
@@ -14,7 +14,7 @@
 	<name>springboot-app-rest-api</name>
 	<description>Spring boot application rest api&apos;s</description>
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -35,7 +35,8 @@
 		<dependency>
 			<groupId>mysql</groupId>
 			<artifactId>mysql-connector-java</artifactId>
-			<scope>runtime</scope>
+			<version>5.1.26</version>
+			<!-- <scope>runtime</scope> -->
 		</dependency>
 		<dependency>
 			<groupId>org.projectlombok</groupId>
@@ -76,6 +77,11 @@
 			<artifactId>springfox-boot-starter</artifactId>
 			<version>3.0.0</version>
 		</dependency>
+	 	<dependency>
+	      	<groupId>io.springfox</groupId>
+	      	<artifactId>springfox-swagger-ui</artifactId>
+	      	<version>3.0.0</version>
+	    </dependency>
 		<dependency>
     		<groupId>org.postgresql</groupId>
 	    	<artifactId>postgresql</artifactId>
@@ -98,7 +104,16 @@
             <artifactId>jaxb-impl</artifactId>
             <version>2.3.1</version>
         </dependency>
-
+        
+        
+        
+ <!--         open API dependency  -->
+<!-- https://mvnrepository.com/artifact/org.springdoc/springdoc-openapi-starter-webmvc-ui -->
+<dependency>
+    <groupId>org.springdoc</groupId>
+    <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+    <version>2.0.2</version>
+</dependency>
 	</dependencies>
 
 	<build>
@@ -113,6 +128,15 @@
 							<artifactId>lombok</artifactId>
 						</exclude>
 					</excludes>
+					<!-- <release>${maven.compiler.release}</release> -->
+				</configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.0</version>
+				<configuration>
+					<release>${maven.compiler.release}</release>
 				</configuration>
 			</plugin>
 		</plugins>
