@@ -4,11 +4,12 @@ import com.springboot.app.payload.CommentDto;
 import com.springboot.app.service.CommentService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import jakarta.validation.Valid;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.List;
 
 @Api(value = "CRUD REST APIs for Comment Respource")
